DefeasibleHelpers

Description

Defeasible helpers module

Documentation

antecedant :: Formula -> Formula

The antecedant function returns the antecedant for a given Formula.

combinations :: Int -> [a] -> [[a]]

The combinations function returns the combinatorial subsets of a list of a specified size.

baseRankAux :: KnowledgeBase -> KnowledgeBase -> RankedModels -> RankedModels

The baseRankAux function is the auxiliary function for the baseRank function.

entailsRCAux :: RankedModels -> Formula -> Bool

The entailsRCAux function is the auxiliary function for the entailsRC function.

entailsLCAux :: RankedModels -> Formula -> Bool

The entailsLCAux function is the auxiliary function for the entailsLC function. NOTE: This is a highly experimental function and its correctness has not been validated.