DefeasibleLogic

Description

Defeasible logic module

Documentation

typically :: Formula -> DefeasibleImplication

The typically function takes in two propositional Formulas and returns a DefeasibleImplication. NOTE: This is a temporary function until a parser is built to handle this

materialise :: DefeasibleKnowledgeBase -> KnowledgeBase

The materialise function returns the corresponding KnowledgeBase for the given DefeasibleKnowledgeBase.

baseRank :: JointKnowledgeBase -> RankedStates

The baseRank function returns the RankedStates for the given JointKnowledgeBase.

entailsRC :: JointKnowledgeBase -> DefeasibleImplication -> Bool

The entailsRC function returns whether or not the given DefeasibleInterpretation is defeasibly entailed (using Rational Closure) by the given JointKnowledgeBase.

entailsLC :: JointKnowledgeBase -> DefeasibleImplication -> Bool

The entailsLC function returns whether or not the given DefeasibleInterpretation is defeasibly entailed (using Lexicographic Closure) by the given JointKnowledgeBase. NOTE: This is a highly experimental function and its correctness has not been validated.