PropositionalHelpers

Description

Propositional helpers module

Documentation

int2bool

Arguments

:: Int
The decimal to convert to binary.

-> Int
The size of the list to be returned (padded with Falses).

-> [Bool]

The int2bool function takes in two ints and returns a Boolean list.

removeDuplicates :: Eq a => [a] -> [a]

The removeDuplicates function returns the given list without duplicates.

subValuation :: [Valuation] -> Valuation -> Bool

The subValuation function returns whether or not at least one of the given list of Valuations is a subset of the given Valuation.

subsumes :: [Valuation] -> [Valuation] -> Bool

The subsumes function takes in two Valuation lists and returns whether the first list subsumes the second.