PropositionalLogic

Description

Propositional logic module

Documentation

atoms :: Formula -> [Name]

The atoms function returns an Atom Name list for the given Formula.

assigns :: Valuation -> Name -> Bool

The assigns function returns the corresponding Atom 's Boolean value for the given Valuation.

satisfies :: Valuation -> Formula -> Bool

The satisfies function evaluates the given Formula for the given Valuation.

valuations :: KnowledgeBase -> [Valuation]

The valuations function returns all possible Valuations for the given KnowledgeBase.

isValid :: Formula -> Bool

The isValid function returns the validity of the given Formula.

models :: KnowledgeBase -> [Valuation]

The models function returns the models (satisfying Valuations) of the given Formula.

entails :: KnowledgeBase -> Formula -> Bool

The entails function returns whether or not the given KnowledgeBase entails the given Formula.

str2form :: String -> Formula

The str2form function returns the Formula encoding for the given String.